feat: add authentication and authorization and also refactor the controllers and classes

+++ Controllers/ToDoListsController.cs
@@ -1,54 +1,97 @@
-﻿using Microsoft.AspNetCore.Mvc;
-using Microsoft.EntityFrameworkCore;
-using HelloToAsp.Data;
+﻿using AutoMapper;
 using HelloToAsp.Contracts;
+using HelloToAsp.Data;
+using HelloToAsp.Dtos.ToDoList;
+using HelloToAsp.Extensions;
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Mvc;
+using Microsoft.EntityFrameworkCore;
 
 namespace HelloToAsp.Controllers
 {
     [Route("api/[controller]")]
     [ApiController]
+    [Authorize(Roles = "SuperAdmin,Admin")]
     public class ToDoListsController : ControllerBase
     {
         private readonly IToDoListRepository _toDoListRepository;
+        private readonly IAuthorizationService _authorizationService;
+        private readonly IMapper _mapper;
 
-        public ToDoListsController(IToDoListRepository toDoListRepository)
+        public ToDoListsController(
+            IToDoListRepository toDoListRepository,
+            IAuthorizationService authorizationService,
+            IMapper mapper
+            )
         {
             _toDoListRepository = toDoListRepository;
+            _authorizationService = authorizationService;
+            _mapper = mapper;
         }
 
         // GET: api/ToDoLists
         [HttpGet]
-        public async Task<ActionResult<IEnumerable<ToDoList>>> GetToDoLists()
+        public async Task<ActionResult<IEnumerable<ToDoListDto>>> GetToDoLists()
         {
-            return await _toDoListRepository.GetAllAsync();
+            var userId = User.GetAuthUserId();
+
+            var data = await _toDoListRepository.GetAllAsync(userId);
+
+            return Ok(data);
         }
 
         // GET: api/ToDoLists/5
         [HttpGet("{id}")]
-        public async Task<ActionResult<ToDoList>> GetToDoList(int id)
+        public async Task<ActionResult<ToDoListGetDetailsDto>> GetToDoList(int id)
         {
-            var toDoList = await _toDoListRepository.GetAsync(id);
+            var userId = User.GetAuthUserId(); // fetch auth id globally
+
+            var toDoList = await _toDoListRepository.GetAsync(userId, id);
 
             if (toDoList == null)
             {
                 return NotFound();
             }
 
-            return toDoList;
+            var authResult = await _authorizationService.AuthorizeAsync(User, toDoList, "ToDoListOwner");
+
+            if (!authResult.Succeeded)
+            {
+                return Forbid();
+            }
+
+            var mappedToDoList = _mapper.Map<ToDoListGetDetailsDto>(toDoList);
+
+            return mappedToDoList;
         }
 
         // PUT: api/ToDoLists/5
         [HttpPut("{id}")]
-        public async Task<IActionResult> PutToDoList(int id, ToDoList toDoList)
+        public async Task<IActionResult> PutToDoList(int id, ToDoListUpdateDto toDoList)
         {
             if (id != toDoList.Id)
             {
                 return BadRequest();
             }
 
+            var existingTask = await _toDoListRepository.GetAsync(id);
+            if (existingTask == null)
+            {
+                return NotFound();
+            }
+
+            var authResult = await _authorizationService.AuthorizeAsync(User, existingTask, "ToDoListOwner");
+
+            if (!authResult.Succeeded)
+            {
+                return Forbid();
+            }
+
+            _mapper.Map(toDoList, existingTask);
+
             try
             {
-                await _toDoListRepository.UpdateAsync(toDoList);
+                await _toDoListRepository.UpdateAsync(existingTask);
             }
             catch (DbUpdateConcurrencyException)
             {
@@ -67,11 +110,15 @@ namespace HelloToAsp.Controllers
 
         // POST: api/ToDoLists
         [HttpPost]
-        public async Task<ActionResult<ToDoList>> PostToDoList(ToDoList toDoList)
+        public async Task<ActionResult> PostToDoList([FromForm] ToDoListCreateDto toDoList)
         {
-            _toDoListRepository.AddAsync(toDoList);
+            var newTask = _mapper.Map<ToDoList>(toDoList);
 
-            return CreatedAtAction("GetToDoList", new { id = toDoList.Id }, toDoList);
+            newTask.UserId = User.GetAuthUserId();
+
+            await _toDoListRepository.AddAsync(newTask);
+
+            return NoContent();
         }
 
         // DELETE: api/ToDoLists/5
@@ -82,6 +129,13 @@ namespace HelloToAsp.Controllers
             if (toDoList == null)
             {
                 return NotFound();
+            }
+
+            var authResult = await _authorizationService.AuthorizeAsync(User, toDoList, "ToDoListOwner");
+
+            if (!authResult.Succeeded)
+            {
+                return Forbid();
             }
 
             await _toDoListRepository.DeleteAsync(id);
